+++ projects/angular8-yandex-maps/src/lib/components/yandex-placemark-component/yandex-placemark.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 
+interface IPlacemark {
+  [key: string]: any;
+}
+
 @Component({
   selector: 'angular-yandex-placemark',
   templateUrl: './yandex-placemark.component.html',
@@ -21,5 +25,12 @@ export class YandexPlacemarkComponent implements OnInit {
       console.error('Placemark: geometry input is required.');
       this.geometry = [];
     }
+  }
+
+  public initPlacemark(ymaps: any, map: any): IPlacemark {
+    const placemark = new ymaps.Placemark(this.geometry, this.properties, this.options);
+
+    map.geoObjects.add(placemark);
+    return placemark;
   }
 }