+++ projects/angular8-yandex-maps/src/lib/services/yandex-map/yandex-service.type.ts
@@ -1,10 +1,5 @@
 import { Subject } from 'rxjs';
 
 export interface IYandexMapService {
-  initMap(): Subject<boolean>;
-  createMap(mapId: string, state: any, options: any): void;
-  addPlacemark(geometry: any, properties: any, options: any): void;
-  addMultiroute(model: any, options: any): void;
-  addGeoObject(feature: any, options: any): void;
-  addSearchControl(request: string, properties: any): void
+  initScript(): Subject<any>;
 }