+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes from v1.0.0 will be documented in this file.
 
+## [1.5.0] - 2019-11-15
+### Added
+- Clusterer for Placemark in Map inputs
+- Examples in README
+
+### Changed
+- Inputs of Map
+
 ## [1.4.0] - 2019-11-09
 ### Changed
 - Inputs of Map
@@ -43,6 +51,7 @@ All notable changes from v1.0.0 will be documented in this file.
 ## [1.0.0] - 2019-09-14
 - Release without changes
 
+[1.5.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.2.0...v1.2.1